+++ apecs/src/Apecs/Components.hs
@@ -1,3 +1,5 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
+
 {-# LANGUAGE ConstraintKinds            #-}
 {-# LANGUAGE FlexibleContexts           #-}
 {-# LANGUAGE FlexibleInstances          #-}
@@ -8,8 +10,6 @@
 {-# LANGUAGE StandaloneDeriving         #-}
 {-# LANGUAGE TemplateHaskell            #-}
 {-# LANGUAGE TypeFamilies               #-}
-
-{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Apecs.Components where
 
@@ -201,18 +201,3 @@ instance Monad m => ExplGet m EntityStore where
   explGet _ ety = return $ Entity ety
   {-# INLINE explExists #-}
   explExists _ _ = return True
-
--- | Pseudocomponent that when written to, actually writes 'c' to its entity argument.
---   Used to dereference during a @cmap@.
-data Redirect c = Redirect Entity c deriving (Eq, Show)
-instance Component c => Component (Redirect c) where
-  type Storage (Redirect c) = RedirectStore (Storage c)
-
-newtype RedirectStore s = RedirectStore s
-type instance Elem (RedirectStore s) = Redirect (Elem s)
-
-instance Has w m c => Has w m (Redirect c) where
-  getStore = RedirectStore <$> getStore
-
-instance (ExplSet m s) => ExplSet m (RedirectStore s) where
-  explSet (RedirectStore s) _ (Redirect (Entity ety) c) = explSet s ety c
